Struct libfprint_rs::FpDevice

#[repr(transparent)]
pub struct FpDevice { /* private fields */ }

Fingerpint device routines. You can interact with fingerprint devices using this struct.

Examples:

use libfprint_rs::FpContext;

let context = FpContext::new();
let devices = context.devices();
let device = devices.get(0).unwrap();

device.open_sync(None).unwrap();
let name = device.name().unwrap();
println!("Device name: {}", name);

Implementations

impl FpDevice

pub fn driver(&self) -> String

The ID of the driver.

pub fn device_id(&self) -> String

The ID of the device.

pub fn name(&self) -> String

The human readable name of the device.

pub fn scan_type(&self) -> FpScanType

Retrieves the scan type of the device.

pub fn nr_enroll_stage(&self) -> i32

Retrieves the number of enroll stages for this device.

pub fn finger_status(&self) -> FpFingerStatus

Retrieves the finger status flags for the device. This can be used by the UI to present the relevant feedback, although it is not guaranteed to be a relevant value when not performing any action.

pub fn features(&self) -> Vec<FpDeviceFeature>

Gets the FpDeviceFeature’s supported by the device .

pub fn has_feature(&self, feature: FpDeviceFeature) -> bool

Checks if device supports the requested FpDeviceFeature.

pub fn is_open(&self) -> bool

Whether the device is open or not

impl FpDevice

pub fn open_sync(&self, cancellable: Option<&Cancellable>) -> Result<(), GError>

Open the device synchronously.

Example:

use libfprint_rs::{FpDevice, FpContext};

let ctx = FpContext::new();
let devices = ctx.devices();
let dev = devices.get(0).unwrap();

dev.open_sync(None).unwrap();

pub fn close_sync( &self, cancellable: Option<&Cancellable> ) -> Result<(), GError>

Close the device synchronously.

Example:

use libfprint_rs::{FpDevice, FpContext};

let ctx = FpContext::new();
let devices = ctx.devices();
let dev = devices.get(0).unwrap();

dev.open_sync(None).unwrap();
dev.close_sync(None).unwrap();

pub fn enroll_sync<T>( &self, template: FpPrint, cancellable: Option<&Cancellable>, progress_cb: Option<FpEnrollProgress<T>>, progress_data: Option<T> ) -> Result<FpPrint, GError>

Enroll a new print. Enrolls a print, progress_cb will be called for each stage of the enrollment process.

Example:

use libfprint_rs::{FpDevice, FpContext, FpPrint};

pub fn enroll_cb(device: &FpDevice,enroll_stage: i32, print: Option<FpPrint>, error: Option<glib::Error>, data: &Option<i32>,) -> () {
    println!("Enroll stage: {}", enroll_stage);
}

let ctx = FpContext::new();
let devices = ctx.devices();
let dev = devices.get(0).unwrap();
dev.open_sync(None).unwrap();

let template = FpPrint::new(&dev);
let new_print = dev.enroll_sync(template, None, Some(enroll_cb), Some(10)).unwrap();

dev.close_sync(None).unwrap();

pub fn verify_sync<T>( &self, enrolled_print: &FpPrint, cancellable: Option<Cancellable>, match_cb: Option<FpMatchCb<T>>, match_data: Option<T>, print: Option<&mut FpPrint> ) -> Result<bool, GError>

Verify a given print synchronously. match_cb will be called when the verification is done.

Example:

use libfprint_rs::{FpDevice, FpContext, FpPrint, GError};

pub fn match_cb(device: &FpDevice, matched_print: Option<FpPrint>, enrolled_print: FpPrint,
error: Option<GError>, data: &Option<i32>) {
    if matched_print.is_some() {
        println!("Matched print: {:?}", matched_print);
    }
}
let ctx = FpContext::new();
let devices = ctx.devices();
let dev = devices.get(0).unwrap();
dev.open_sync(None).unwrap();

let some_print: FpPrint = foreign_function_that_gets_print();
let mut new_print = FpPrint::new(&dev); // The variable that will hold the new print
let verified = dev.verify_sync(&some_print, None, Some(match_cb), Some(10), Some(&mut
new_print)).unwrap();
if verified {
   println!("Print verified");println
}

pub fn suspend_sync( &self, cancellable: Option<&Cancellable> ) -> Result<(), GError>

Prepare device for suspend.

pub fn resume_sync( &self, cancellable: Option<&Cancellable> ) -> Result<(), GError>

Resume device after suspend.

pub fn identify_sync<T>( &self, prints: &Vec<FpPrint>, cancellable: Option<&Cancellable>, match_cb: Option<FpMatchCb<T>>, match_data: Option<T>, print: Option<&mut FpPrint> ) -> Result<Option<FpPrint>, GError>

Identify a print synchronously.

match_cb will be called when a print matches or at the end of the operation.

Example:

use libfprint_rs::{FpDevice, FpContext, FpPrint, GError};

pub fn match_cb(device: &FpDevice, matched_print: Option<FpPrint>, enrolled_print: FpPrint,
error: Option<GError>, data: &Option<i32>) {
    if matched_print.is_some() {
        println!("Matched print: {:?}", matched_print);
    }
}

let ctx = FpContext::new();
let devices = ctx.devices();
let dev = devices.get(0).unwrap();
dev.open_sync(None).unwrap();

let vec_prints: Vec<FpPrint> = function_returning_Vec_prints();
let mut new_print = FpPrint::new(&dev); // The variable that will hold the new print
let print_identified = dev.identify_sync(&vec_prints, None, Some(match_cb), Some(10), Some(&mut
new_print)).unwrap();
if print_identified.is_some() {
    println!("Found matching print on vector passed");
}

pub fn capture_sync( &self, wait_for_finger: bool, cancellable: Option<&Cancellable> ) -> Result<FpImage, GError>

Start an synchronous operation to capture an image.

Example:

let ctx = FpContext::new();
let devices = ctx.devices();
let dev = devices.get(0).unwrap();
dev.open_sync(None).unwrap();

let image = dev.capture_sync(true, None).unwrap();

pub fn delete_print_sync()

Delete a given print from the device.

pub fn list_prints_sync()

List device stored prints synchronously.

pub fn clear_storage_sync()

Clear sensor storage.