Struct libfprint_rs::FpPrint

#[repr(transparent)]
pub struct FpPrint { /* private fields */ }

Struct representing a fingerprint.

Implementations

impl FpPrint

pub fn new(dev: &FpDevice) -> Self

Create a new FpPrint. This is only useful to prepare an enrollment of a new print using FpDevice::enroll_sync. For this you should first create a new print, fill in the relevant metadata, and then start the enrollment

pub fn driver(&self) -> String

Returns the driver that the print was created for.

pub fn device_id(&self) -> String

Returns the device ID that the print was created for.

pub fn device_stored(&self) -> bool

Whether the print is actually stored on the device and this is just a handle to use that references the device stored data.

pub fn image(&self) -> Option<FpImage>

Returns the image that the print was created from, or None

pub fn finger(&self) -> FpFinger

Returns the finger that the print was created for.

pub fn username(&self) -> Option<String>

Returns the user defined username for the print.

pub fn description(&self) -> Option<String>

Returns the user defined description for the print.

pub fn enroll_date(&self) -> Option<GDate>

Returns the user defined enroll date for the print.

pub fn set_finger(&self, finger: FpFinger)

Set the finger that the print is for.

pub fn set_username(&self, username: &str)

Set the username for the print.

pub fn set_description(&self, description: &str)

Set the description for the print.

pub fn set_enroll_date(&self, enroll_date: GDate)

Set the enroll date for the print.

pub fn compatible(&self, device: &FpDevice) -> bool

Tests whether the prints is compatible with the given device.

pub fn equal(&self, other: &FpPrint) -> bool

Tests whether the prints can be considered equal. This only compares the actual information about the print, not the metadata.

pub fn serialize(&self) -> Result<Vec<u8>, Error>

Serialize a print definition for permanent storage. Note that this is lossy in the sense that e.g. the image data is discarded.

pub fn deserialize(data: &[u8]) -> Result<FpPrint, Error>

Deserialize a print definition from permanent storage.