Type Definition libfprint_rs::FpMatchCb

pub type FpMatchCb<T> = fn(_: &FpDevice, _: Option<FpPrint>, _: FpPrint, _: Option<GError>, _: &Option<T>);

This type represents the callback function for the FpDevice::verify and FpDevice::identify implementations and will be called when a print is matched.